small bug fix with drawing function

+++ CoordinateSystem.h
@@ -17,7 +17,7 @@ public:
     double scale;
     float length_of_step;
     double pixel_step;
-    int start_cart_x, end_cart_x,start_cart_y, end_cart_y;
+    double start_cart_x, end_cart_x,start_cart_y, end_cart_y;
     CoordinateSystem(int screen_width, int screen_height,float scale,int gridStep);
     void DrawGrid();
     void DrawCoordinateSystem();
